code refactoring, #140
