+++ src/main/java/org/metanorma/fop/PDFGenerator.java
@@ -54,6 +54,7 @@ import static org.metanorma.Constants.*;
 import static org.metanorma.fop.fontConfig.DEFAULT_FONT_PATH;
 import static org.metanorma.fop.Util.getStreamFromResources;
 
+import org.metanorma.fop.eventlistener.SecondPassSysOutEventListener;
 import org.metanorma.fop.ifhandler.FOPIFIndexHandler;
 import org.metanorma.utils.LoggerHelper;
 import org.xml.sax.InputSource;
@@ -349,9 +350,8 @@ public class PDFGenerator {
     /**
      * Converts an XML file to a PDF file using FOP
      *
-     * @param config the FOP config file
-     * @param xml the XML source file
-     * @param xsl the XSL file
+     * @param fontcfg the FOP config file
+     * @param xsltConverter the XSL converter
      * @param pdf the target PDF file
      * @throws IOException In case of an I/O problem
      * @throws FOPException, SAXException In case of a FOP problem
@@ -911,54 +911,7 @@ public class PDFGenerator {
 
     }*/
 
-    /** A simple event listener that writes the events to stdout and sterr. */
-    //private static class SysOutEventListener implements org.apache.fop.events.EventListener {
 
-        /** {@inheritDoc} */
-   /*     public void processEvent(Event event) {
-            String msg = EventFormatter.format(event);
-            EventSeverity severity = event.getSeverity();
-            if (severity == EventSeverity.INFO) {
-                System.out.println("[INFO ] " + msg);
-            } else if (severity == EventSeverity.WARN) {
-                System.out.println("[WARN ] " + msg);
-            } else if (severity == EventSeverity.ERROR) {
-                System.err.println("[ERROR] " + msg);
-            } else if (severity == EventSeverity.FATAL) {
-                System.err.println("[FATAL] " + msg);
-            } else {
-                assert false;
-            }
-        }
-    }*/
-    
-    /** A simple event listener that writes the events to stdout and sterr. */
-    private static class SecondPassSysOutEventListener implements org.apache.fop.events.EventListener {
-
-        /** {@inheritDoc} */
-        public void processEvent(Event event) {
-            String msg = EventFormatter.format(event);
-            EventSeverity severity = event.getSeverity();
-            if (severity == EventSeverity.INFO) {
-                if(msg.startsWith("Rendered page #")) {
-                    //System.out.println("[INFO] Intermediate format. " + msg);
-                    logger.log(Level.INFO, "[INFO] Intermediate format. {0}", msg);
-                }
-            } else if (severity == EventSeverity.WARN) {
-                //System.out.println("[WARN] " + msg);
-            } else if (severity == EventSeverity.ERROR) {
-                //System.err.println("[ERROR] " + msg);
-                logger.log(Level.SEVERE, "[ERROR] {0}", msg);
-            } else if (severity == EventSeverity.FATAL) {
-                //System.err.println("[FATAL] " + msg);
-                logger.log(Level.SEVERE, "[FATAL] {0}", msg);
-            } else {
-                assert false;
-            }
-        }
-    }
-    
-    
     private int getPageCount() {
         return pageCount;
     }
